Moved dateToKey to utils.js
