+++ src/server/index.js
@@ -1,2 +1,4 @@
 export * from './auth.js';
 export * from './appointments.js';
+export * from './records.js';
+export * from './utils.js';
